# Client response muxer