… README.md
… protobuf-nats.gemspec
… spec/fake_nats_client.rb
… spec/protobuf/nats/jnats_spec.rb
… spec/spec_helper.rb